Fix installing helm chart on AWS EKS not working
AWS EKS is using an unconventional semantic versioning scheme: `major.minor.patch-metadata`. Helm is expecting something like `major.minor.patch[-prerelease][+metadata]`.

This causes helm's semver check to interpret an EKS version such as `v1.19.13-eks-8df270` as a pre-release, which causes it not satisfy a constraint such as `>= 1.18.0`.

More information in this issue: https://github.com/helm/helm/issues/10375

With neither AWS or Helm willing to change anything on their end, in order to use EKS and Helm one must set the constraint such that it allows pre-releases. This change allows any release or pre-release of Kubernetes with a version greater or equal to 1.18.0.